pyproject.toml
[tool.poetry]
name = "batch"
version = "0.1.0"
description = "에테르 배치"
authors = ["이재우 <jwleeitn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python ="^3.10"
aiohappyeyeballs="^2.4.3"
aiohttp="^3.10.10"
aiosignal="^1.3.1"
annotated-types="^0.7.0"
anyio="^4.6.2.post1"
async-timeout="^5.0.1"
asyncpg="^0.29.0"
attrs="^24.2.0"
beautifulsoup4="^4.12.3"
certifi="^2024.8.30"
charset-normalizer="^3.4.0"
click="^8.1.7"
colorama="^0.4.6"
common={git="https://github.com/ether-trading-system/common.git"}
databases="^0.9.0"
fastapi="^0.112.4"
frozenlist="^1.5.0"
greenlet="^3.1.1"
h11="^0.14.0"
idna="^3.10"
multidict="^6.1.0"
numpy="^2.1.3"
pandas="^2.2.3"
propcache="^0.2.0"
psycopg2="^2.9.10"
pydantic="^2.9.2"
pydantic-settings="^2.6.1"
pydantic_core="^2.23.4"
PyMySQL="^1.1.1"
python-dateutil="^2.9.0.post0"
python-dotenv="^1.0.1"
pytz="^2024.2"
PyYAML="^6.0.2"
requests="^2.32.3"
six="^1.16.0"
sniffio="^1.3.1"
soupsieve="^2.6"
SQLAlchemy="^2.0.36"
starlette="^0.38.6"
typing_extensions="^4.12.2"
tzdata="^2024.2"
urllib3="^2.2.3"
uvicorn="^0.31.1"
yarl="^1.17.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
